+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Service for order dispatch in `src/api/v1/apps/services.py`
+- Service for waiter dashboard in `src/api/v1/apps/services.py`
+- Templates for waiter dashboard:
+  - Dashboard `src/templates/orders/waiter/dashboard.html`
+  - Cards `src/templates/orders/waiter/cards.html`
 
 ### Changed
 - Improve the UI design, changes names and delete prints for debugging in `src/templates/orders/kitchen/dashboard.html`